requirements.txt
CrewAI
opencv-python
crewai-tools
langchain_openai
langchain
python-dotenv